Corrijo: - Problemas en los módulos de la base de datos relacionados con las tuplas vacías. - Problema al cargar un script SQL en el configurador de la base de datos.
 On branch vmserver-integration
 Changes to be committed:

	modified:   src/CygnusCloud/database/MainServerDB.sql
	modified:   src/CygnusCloud/database/mainServer/mainServerDB.py
	modified:   src/CygnusCloud/database/utils/configuration.py
	modified:   src/CygnusCloud/database/vmServer/imageManager.py
	modified:   src/CygnusCloud/database/vmServer/runtimeData.py
	modified:   src/CygnusCloud/mainServer/mainServer.py
	modified:   src/CygnusCloud/mainServer/reactor.py
	modified:   src/CygnusCloud/utils/commands.py
	modified:   src/CygnusCloud/virtualMachineServer/constantes.py
	modified:   src/CygnusCloud/virtualNetwork/virtualNetworkManager.py

+++ src/CygnusCloud/database/vmServer/runtimeData.py
@@ -28,7 +28,7 @@ class RuntimeData(BasicDatabaseConnector):
         #Ejecutamos el comando
         self._executeUpdate(sql)  
         #Creamos la tabla necesaria
-        sql = "CREATE TABLE IF NOT EXISTS freeMacs(UUID VARCHAR(40) ,MAC VARCHAR(20),PRIMARY KEY(UUID,MAC))"
+        sql = "CREATE TABLE IF NOT EXISTS freeMacs(UUID VARCHAR(40) ,MAC VARCHAR(20),PRIMARY KEY(UUID,MAC)) ENGINE=MEMORY;"
         #Ejecutamos el comando
         self._executeUpdate(sql)
         #Generamos el relleno
@@ -51,11 +51,11 @@ class RuntimeData(BasicDatabaseConnector):
         '''
             Función encargada de crear la tabla inicial de pares (UUID,MAC) libres
         '''
-        sql = "DROP TABLE IF EXISTS freeVNCPorts" 
+        sql = "DROP TABLE IF EXISTS freeVNCPorts;" 
         #Ejecutamos el comando
         self._executeUpdate(sql)  
         #Creamos la tabla necesaria
-        sql = "CREATE TABLE IF NOT EXISTS freeVNCPorts(VNCPort INTEGER PRIMARY KEY)"
+        sql = "CREATE TABLE IF NOT EXISTS freeVNCPorts(VNCPort INTEGER PRIMARY KEY) ENGINE=MEMORY;"
         #Ejecutamos el comando
         self._executeUpdate(sql)
         #Generamos el relleno
@@ -80,7 +80,8 @@ class RuntimeData(BasicDatabaseConnector):
         sql = "SELECT * FROM freeMacs"
         #Nos quedamos con la primera ocurrencia
         result=self._executeQuery(sql, True)
-        
+        if (result == ()) : 
+            return None        
         #Eliminamos este resultado de la tabla
         sql = "DELETE FROM freeMacs WHERE UUID = '" + result[0] + "' AND MAC ='" + result[1] + "'"
         #Ejecutamos el comando
@@ -109,7 +110,8 @@ class RuntimeData(BasicDatabaseConnector):
         sql = "SELECT * FROM freeVNCPorts"
         #Nos quedamos con la primera ocurrencia
         result=self._executeQuery(sql, True)
-        
+        if (result == ()) : 
+            return None
         #Eliminamos este resultado de la tabla
         sql = "DELETE FROM freeVNCPorts WHERE VNCPort = '" + str(result[0]) + "'"
         #Ejecutamos el comando
@@ -184,6 +186,8 @@ class RuntimeData(BasicDatabaseConnector):
         sql = "SELECT VMId FROM ActualVM WHERE VNCPortAdress = '" + str(vncPort) + "'" 
         #Recogemos el resultado
         result=self._executeQuery(sql, True)
+        if (result == ()) : 
+            return None
         #Devolvemos el resultado
         return result[0] 
     
@@ -196,6 +200,8 @@ class RuntimeData(BasicDatabaseConnector):
         sql = "SELECT VMId FROM ActualVM WHERE domainName = '" + str(domainName) + "'" 
         #Recogemos el resultado
         result=self._executeQuery(sql, True)
+        if (result == ()) : 
+            return None
         #Devolvemos el resultado
         return result[0] 
     
@@ -209,6 +215,8 @@ class RuntimeData(BasicDatabaseConnector):
         sql += "vm.VMId = av.VMId " 
         #Recogemos el resultado
         result=self._executeQuery(sql, True)
+        if (result == ()) : 
+            return None
         #Devolvemos el resultado
         return result[0] 
     
@@ -222,6 +230,8 @@ class RuntimeData(BasicDatabaseConnector):
         sql += "vm.VMId = av.VMId " 
         #Recogemos el resultado
         result=self._executeQuery(sql, True)
+        if (result == ()) : 
+            return None
         #Devolvemos el resultado
         return result[0] 
     
@@ -234,6 +244,8 @@ class RuntimeData(BasicDatabaseConnector):
         sql = "SELECT imageCopyPath FROM ActualVM  WHERE VNCPortAdress = '" + str(vncPort) + "'"
         #Recogemos el resultado
         result=self._executeQuery(sql, True)
+        if (result == ()) : 
+            return None
         #Devolvemos el resultado
         return result[0] 
     
@@ -246,6 +258,8 @@ class RuntimeData(BasicDatabaseConnector):
         sql = "SELECT imageCopyPath FROM ActualVM  WHERE domainName = '" + str(domainName) + "'"
         #Recogemos el resultado
         result=self._executeQuery(sql, True)
+        if (result == ()) : 
+            return None
         #Devolvemos el resultado
         return result[0] 
     
@@ -257,6 +271,8 @@ class RuntimeData(BasicDatabaseConnector):
         sql = "SELECT macAdress FROM ActualVM WHERE VNCPortAdress = '" + str(vncPort) + "'" 
         #Recogemos el resultado
         result=self._executeQuery(sql, True)
+        if (result == ()) : 
+            return None
         #Devolvemos el resultado
         return result[0] 
     
@@ -268,6 +284,8 @@ class RuntimeData(BasicDatabaseConnector):
         sql = "SELECT macAdress FROM ActualVM WHERE domainName = '" + str(domainName) + "'" 
         #Recogemos el resultado
         result=self._executeQuery(sql, True)
+        if (result == ()) : 
+            return None
         #Devolvemos el resultado
         return result[0] 
     
@@ -279,6 +297,8 @@ class RuntimeData(BasicDatabaseConnector):
         sql = "SELECT uuid FROM ActualVM WHERE VNCPortAdress = '" + str(vncPort) + "'" 
         #Recogemos el resultado
         result=self._executeQuery(sql, True)
+        if (result == ()) : 
+            return None
         #Devolvemos el resultado
         return result[0] 
     
@@ -290,6 +310,8 @@ class RuntimeData(BasicDatabaseConnector):
         sql = "SELECT uuid FROM ActualVM WHERE domainName = '" + str(domainName) + "'" 
         #Recogemos el resultado
         result=self._executeQuery(sql, True)
+        if (result == ()) : 
+            return None
         #Devolvemos el resultado
         return result[0] 
     
@@ -301,6 +323,8 @@ class RuntimeData(BasicDatabaseConnector):
         sql = "SELECT VNCPass FROM ActualVM WHERE VNCPortAdress = '" + str(vncPort) + "'" 
         #Recogemos el resultado
         result=self._executeQuery(sql, True)
+        if (result == ()) : 
+            return None
         #Devolvemos el resultado
         return result[0]
     
@@ -312,6 +336,8 @@ class RuntimeData(BasicDatabaseConnector):
         sql = "SELECT VNCPass FROM ActualVM WHERE domainName = '" + str(domainName) + "'" 
         #Recogemos el resultado
         result=self._executeQuery(sql, True)
+        if (result == ()) : 
+            return None
         #Devolvemos el resultado
         return result[0]
     
@@ -323,6 +349,8 @@ class RuntimeData(BasicDatabaseConnector):
         sql = "SELECT VMPid FROM ActualVM WHERE VNCPortAdress = '" + str(vncPort) + "'" 
         #Recogemos el resultado
         result=self._executeQuery(sql, True)
+        if (result == ()) : 
+            return None
         #Devolvemos el resultado
         return result[0]
     
@@ -334,6 +362,8 @@ class RuntimeData(BasicDatabaseConnector):
         sql = "SELECT VMPid FROM ActualVM WHERE domainName = '" + str(domainName) + "'" 
         #Recogemos el resultado
         result=self._executeQuery(sql, True)
+        if (result == ()) : 
+            return None
         #Devolvemos el resultado
         return result[0]
     
@@ -344,7 +374,9 @@ class RuntimeData(BasicDatabaseConnector):
         #Creamos la consulta encargada de extraer los datos
         sql = "SELECT osImagePath FROM ActualVM WHERE VNCPortAdress = '" + str(vncPort) + "'" 
         #Recogemos el resultado
-        result=self._executeQuerY(sql, True)
+        result=self._executeQuery(sql, True)
+        if (result == ()) : 
+            return None
         #Devolvemos el resultado
         return result[0]
     
@@ -356,6 +388,8 @@ class RuntimeData(BasicDatabaseConnector):
         sql = "SELECT osImagePath FROM ActualVM WHERE domainName = '" + str(domainName) + "'" 
         #Recogemos el resultado
         result=self._executeQuery(sql, True)
+        if (result == ()) : 
+            return None
         #Devolvemos el resultado
         return result[0]    
     
@@ -367,6 +401,8 @@ class RuntimeData(BasicDatabaseConnector):
         sql = "SELECT domainName FROM ActualVM WHERE VNCPortAdress = '" + str(vncPort) + "'" 
         #Recogemos el resultado
         result=self._executeQuery(sql, True)
+        if (result == ()) : 
+            return None
         #Devolvemos el resultado
         return result[0]
     
@@ -406,6 +442,8 @@ class RuntimeData(BasicDatabaseConnector):
         sql = "SELECT COUNT(*) FROM ActualVM WHERE VNCPortAdress =" + str(port)
         #Recogemos los resultado
         result=self._executeQuery(sql, True)
+        if (result == ()) : 
+            return None
         #Si el resultado es 1, la MV existirá
         return (result[0] == 1) 
     
@@ -418,11 +456,10 @@ class RuntimeData(BasicDatabaseConnector):
         sql = "SELECT userId FROM ActualVM WHERE domainName = '" + str(domainName) + "'" 
         #Recogemos el resultado
         result=self._executeQuery(sql, True)
+        if (result == ()) : 
+            return None
         #Devolvemos el resultado
-        if result != None :
-            return result[0]
-        else :
-            return result
+        return result[0]
         
 def main():    
     #Instanciamos la clase
